+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 - [SearchOptions] Remove SearchOptions.AttributeSets parameter.
 
+- [SearchResult] Add ResultChildMetadata type to SearchResultMetadata.children field.
+
 - [SearchEngine] Add RetrieveOptions(attributeSets:) parameter to `select()` functions. Use this to provide attribute set queries.
 
 **MapboxCommon**: v24.6.0